Partially address comments from review
