+++ types/src/front/vault.ts
@@ -1,5 +1,3 @@
-import { GroupType } from "./groups";
-
 const VAULT_KINDS = ["regular", "global", "system", "public"] as const;
 export type VaultKind = (typeof VAULT_KINDS)[number];
 
@@ -8,8 +6,4 @@ export type VaultType = {
   sId: string;
   kind: VaultKind;
   groupIds: string[];
-};
-
-export type PokeVaultType = VaultType & {
-  groups: GroupType[];
 };
